add src/getPhotosByAlbumId.js, test/getPhotosByAlbumId.test.js
Using Stubs for Testing in JavaScript with Sinon.js
https://stackabuse.com/using-stubs-for-testing-in-javascript-with-sinon-js/

https://sinonjs.org/releases/latest/stubs/
stub.yields([arg1, arg2, ...])
Similar to callsArg.
Causes the stub to call the first callback it receives with the provided arguments (if any).
If a method accepts more than one callback, you need to use yieldsRight to call the last callback or callsArg to have the stub invoke other callbacks than the first or last one.

https://www.coder.work/article/2595822
javascript - sinon的stub.yields有什么作用？
如果您要存根的函数接受回调，例如异步数据库请求，这将允许存根伪造该函数通常传递给您的回调的结果。